feat:搭建订单服务的基本框架

+++ order-service/src/main/java/com/sdu/rabbitmq/order/entity/vo/CreateOrderVO.java
@@ -8,7 +8,10 @@ import lombok.ToString;
 @Setter
 @ToString
 public class CreateOrderVO {
-    private Integer accountId;
+
+    private Long accountId;
+
     private String address;
-    private Integer productId;
+
+    private Long productId;
 }
